fix: proper Trace ID comparison (#70)
Previously, Trace IDs were being compared using the `~=` (not equal)
operator, which is not suitable for Lua tables. This led to incorrect
evaluation, resulting in false positive. As a consequence, logs were
unnecessarily spammed when a request contained mutliple tracing
contexts.

This commit resolves the issue by implementing a proper table comparison
for Trace IDs.